feat: support `--list-packages` command flag
Currently, this is happening _after_ we load the database which is technically not required for
parsing, but I think it makes sense to load the database first on the main flow so have left it as
is for now.

We should improve on this in the future - I'm hoping we can load the database in parallel to
parsing, but even just having a "don't need to load" check would be fine.